--- crates/oxc_traverse/tests/compile_fail.rs
@@ -1,5 +1,0 @@
-#[test]
-fn compile_fail() {
-    let t = trybuild::TestCases::new();
-    t.compile_fail("tests/compile_fail/*.rs");
-}
