--- crates/oxc_traverse/tests/compile_fail/ancestor_lifetime1.stderr
@@ -1,11 +1,0 @@
-error: lifetime may not live long enough
-  --> tests/compile_fail/ancestor_lifetime1.rs:14:9
-   |
-8  | impl<'a, 'b> Traverse<'a> for Trans<'a, 'b> {
-   |          -- lifetime `'b` defined here
-...
-12 |         ctx: &mut TraverseCtx<'a>,
-   |              - let's call the lifetime of this reference `'1`
-13 |     ) {
-14 |         self.ancestor = Some(ctx.parent());
-   |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ assignment requires that `'1` must outlive `'b`
