--- crates/oxc_traverse/tests/compile_fail/ancestor_lifetime3.stderr
@@ -1,11 +1,0 @@
-error: lifetime may not live long enough
-  --> tests/compile_fail/ancestor_lifetime3.rs:16:13
-   |
-9  | impl<'a, 'b> Traverse<'a> for Trans<'a, 'b> {
-   |          -- lifetime `'b` defined here
-...
-13 |         ctx: &mut TraverseCtx<'a>,
-   |              - let's call the lifetime of this reference `'1`
-...
-16 |             self.program_body = Some(program.body());
-   |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ assignment requires that `'1` must outlive `'b`
